refactor: isolate RoboCasa365 benchmark
