+++ configs/benchmarks/robocasa/README.md
@@ -2,7 +2,7 @@
 smoke_config: smoke.yaml
 ---
 
-# RoboCasa and RoboCasa365
+# RoboCasa365
 
 Official 50-task multi-task benchmark using RoboCasa's Panda-Omron Gymnasium wrapper.
 [Paper](https://robocasa.ai/assets/robocasa365_iclr26.pdf) ·
@@ -16,12 +16,10 @@ Official 50-task multi-task benchmark using RoboCasa's Panda-Omron Gymnasium wra
 
 | File | Description | Tasks | Episodes/task |
 | --- | --- | ---: | ---: |
-| `eval.yaml` | Legacy atomic-task evaluation | 6 | 5 |
 | `rc365.yaml` | Official multi-task protocol | 50 | 50 |
 | `smoke.yaml` | Contract smoke test | 2 | 1 |
 
 The adapter reads task membership and task-specific horizons from RoboCasa's registry.
-The `rc365.yaml` and `smoke.yaml` configs select this behavior with `protocol: rc365`; omitting `protocol` preserves the legacy arbitrary-task, configurable-camera, 7-D adapter.
 It evaluates the target50 tasks in pretraining kitchens (`split: pretrain`), matching the official multi-task leaderboard protocol.
 Success is sampled and accumulated at the end of each 16-action chunk, while every episode continues to its task-specific horizon.
 The wire contract preserves all 12 Panda-Omron dimensions in the official dataset order instead of padding a 7-D arm action.
