--- configs/benchmarks/robocasa/eval.yaml
@@ -1,24 +1,0 @@
-# RoboCasa kitchen manipulation evaluation (6 tasks × 5 episodes)
-#
-# RoboCasa provides 365 kitchen tasks on 2500+ procedurally-generated scenes.
-# Default config uses a small subset of atomic tasks for quick evaluation.
-#
-# Requires ~10GB of kitchen assets from HuggingFace.
-server:
-  url: "ws://localhost:8000"
-
-docker:
-  image: ghcr.io/allenai/vla-evaluation-harness/robocasa:latest
-
-output_dir: "./results"
-
-benchmarks:
-  - benchmark: "vla_eval.benchmarks.robocasa.benchmark:RoboCasaBenchmark"
-    episodes_per_task: 5
-    params:
-      robot: PandaOmron
-      camera_size: 256
-      max_steps: 500
-      split: pretrain
-      seed: null
-
