+++ src/vla_eval/benchmarks/robocasa/__init__.py
@@ -1,5 +1,6 @@
-"""RoboCasa kitchen manipulation benchmark."""
+"""RoboCasa and RoboCasa365 kitchen manipulation benchmarks."""
 
 from vla_eval.benchmarks.robocasa.benchmark import RoboCasaBenchmark
+from vla_eval.benchmarks.robocasa.rc365 import RoboCasa365Benchmark
 
-__all__ = ["RoboCasaBenchmark"]
+__all__ = ["RoboCasa365Benchmark", "RoboCasaBenchmark"]
